[ast::parse] add 'Parser'
'Parser' allows exposing all of the parsing methods without importing
the entire 'parse' module.  The methods were rewritten to use 'Parser'
to access storage.  The biggest change was the implementation of a
custom Pratt parser based on calculated precedences - it reports cases
where incompatible operators are chained, resulting in lots of new
'Result' cases in the API.  'main()' once again parses the input.