[mir] document operations
This module represents how 'mir' will work in the future; it will contain
instructions based on these operations, each of which tracks all the necessary
type information.  It also relies on a more semantic definition of types, which
is coming up soon.

The benefit of these definitions is that the LLVM IR for an instruction can be
generated directly given an operation and its inputs.  It cleanly represents the
necessary mapping operations as well.