[hir] overhaul AST->HIR parsing
This is an unprecedently large change, completely re-evaluating how parsing
works.  Everything is now divided into the different type layers, with various
bits of mapping and zipping at each layer.

The only thing left to support is casting.  It will be implemented as a new
unary operation with fairly unusual typing rules.